modified input type

+++ generate_regular_grid.cpp
@@ -16,11 +16,11 @@
 using namespace std;
 
 // Get x values at the grid
-double** generate_x_points (double x_extent, double y_extent, int nx, int ny) {
+int** generate_x_points (int x_extent, int y_extent, int nx, int ny) {
 
-    double** x_grid = (double**) malloc(ny * sizeof(double*));
+    int** x_grid = (int**) malloc(ny * sizeof(int*));
     for (int ind = 0; ind < ny; ind++) {
-        x_grid[ind] = (double*) malloc (nx * sizeof (double));
+        x_grid[ind] = (int*) malloc (nx * sizeof (int));
     }
 
     double dx = get_dx (x_extent, nx); // Get spacing between grid along x
@@ -36,11 +36,11 @@ double** generate_x_points (double x_extent, double y_extent, int nx, int ny) {
 }
 
 //Get Y values at the grid
-double** generate_y_points (double x_extent, double y_extent, int nx, int ny) {
+int** generate_y_points (int x_extent, int y_extent, int nx, int ny) {
 
-    double** y_grid = (double**) malloc (ny * sizeof(double*));
+    int** y_grid = (int**) malloc (ny * sizeof(int*));
     for (int i = 0; i < ny; i++) {
-        y_grid[i] = (double*) malloc (nx * sizeof (double));
+        y_grid[i] = (int*) malloc (nx * sizeof (int));
     }
 
     double dy = get_dy (y_extent, ny); // Get spacing between grid along y
