+++ generate_regular_grid.h
@@ -1,16 +1,16 @@
 #ifndef REGULAR_GRID
 #define REGULAR_GRID
 
-double get_dx (double x_extent, int nx) {
+double get_dx (int x_extent, int nx) {
     return x_extent/nx;
 }
 
-double get_dy (double y_extent, int ny) {
+double get_dy (int y_extent, int ny) {
     return y_extent/ny;
 }
 
-double** generate_x_points (double x_extent, double y_extent, int nx, int ny);
+int** generate_x_points (int x_extent, int y_extent, int nx, int ny);
 
-double** generate_y_points (double x_extent, double y_extent, int nx, int ny);
+int** generate_y_points (int x_extent, int y_extent, int nx, int ny);
 
 #endif
